Fase 2/Evidencias Proyecto/Desarrollo del proyecto/requirements.txt
django-quill-editor
django-crispy-forms
django-livereload-server
pillow
crispy-bootstrap4
cx_oracle
django
colorama
tabulate